requirements.txt
flask
requests
beautifulsoup4
google-generativeai
google-genai
schedule
markdown
pymupdf
reportlab